requirements.txt
langchain
python-dotenv
yfinance
streamlit
langchain_google_genai
langchain_core
langchain_community
PyPDF2
pypdf
pdfplumber
Pillow
pytesseract
opencv-python
reportlab
fpdf2
streamlit-drawable-canvas
pandas
numpy
faiss-cpu
markdown2
weasyprint
PyMuPDF
fitz
streamlit-camera-input-live